requirements.txt
Django>=3.0.6
django-allauth==0.41.0
django-cities-light==3.6.0
django-cors-headers==3.3.0
django-countries==6.1.2
django-crispy-forms==1.9.1
django-phonenumber-field==4.0.0
django-rest-framework==0.1.0
djangorestframework>=3.11.0
django-rest-auth==0.9.5
misaka==2.1.1
phonenumbers==8.12.4
Pillow==7.0.0
six>=1.9.0
sqlparse>=0.2.2
asgiref==3.2.7
pytz>=2019.3
